+++ Assets/Weapons/Blaster/Scripts/PlasmaProjectile.cs
@@ -4,7 +4,7 @@ namespace Weapons
 {
     public class PlasmaProjectile : Projectile
     {
-        public override void Launch(float speed, float? angle = null)
+        public override void Launch(int speed, float? angle = null)
         {
             _rb.AddForce(new Vector2(speed, 0), ForceMode2D.Impulse);
         }
